Add zipJoin() as an alternative to join() for RDDs with sorted partitions.
The benchmark results I see:

      rows |  join (ms) | zipJoin (ms)
   4000000 |       5179 |       1664
   4400000 |       8149 |       1876
   4800000 |       8093 |       2383
   5200000 |       6507 |       1950
   5600000 |       6599 |       2142
   6000000 |      11730 |       2304
   6400000 |       7737 |       2238
   6800000 |       8523 |       7409
   7200000 |      12795 |       2564
   7600000 |      12497 |       2678
   8000000 |      14155 |       2944

Another run: (There's a fair bit of variance.)

      rows |  join (ms) | zipJoin (ms)
   4000000 |       5000 |       1757
   4400000 |       5560 |       1747
   4800000 |       5140 |       1847
   5200000 |       7977 |       2375
   5600000 |       8534 |       2180
   6000000 |      10778 |       2588
   6400000 |       6831 |       4530
   6800000 |      10900 |       2922
   7200000 |       9169 |       5228
   7600000 |       7786 |       2541
   8000000 |      14653 |       3427

It beats join each time, but I see way bigger differences when running
much the same test on a standalone Spark. (Like 30s vs 0.3s.) I am not sure if
my methodology is at fault, or this is a matter of local vs standalone,
or if our Spark setup is way better than the default (e.g. we use Kryo).