Revert "Add a Protractor plugin to wait for the slow requests queue."
This reverts commit 2685b081f5d360ceae6e609171e0d9c4a18a7ce2.

I don't think this is actually needed. I never saw a non-zero-length
queue in this plugin. It is logical too: Protractor waits for $http
requests and for digests. The queue starts one $http request after the
other (perhaps in a digest -- I'm not sure how promises are
implemented). I believe it's enough to make sure Protractor waits for
the entire queue.

Nevertheless we can simply put back this plugin if we see issues around
this.