Get real certificate for "local.lynxkite.com".
I am committing the private part of a certificate along with its password into the GitHub
repository. The idea with this hostname ("local.lynxkite.com") is that it will never be used online.
It is not a subdomain of "lynxanalytics.com". So I think it is reasonable to not worry about
exposing the credentials for this domain.

The only risk posed by a lax handling of this certificate is that someone could use it to
impersonate LynxKite on a company network. So just like the localhost certificate, this certificate
should also be used exclusively in controlled networks. The difference is that it can be easier to
set up.