Replacement for writeAttributes.
The original code was very fancy to avoid reading the input multiple
times. This was a concern, for example with large CSVs. But now the
input is always a Table, which is backed directly by a Parquet file in
LynxKite storage. Parquet is column-oriented, so we can just read the
columns one by one.